Added sample OPS1955 code snippets
The Custom Kapsch OPS1955 ASN.1 specs tests never passed properly using
either PER or OER. Not because the ASN.1 specs were incorrect, but
because the OPS1955 L7 Ethernet data structures did not follow any of
the ASN.1 encoding rules.
First, in OER, the length of a SEQUENCE OF (AttributeList) was encoded
in a single byte. When in actuality, in OER, INTEGER always has a
length prefix itself, and the length of a SEQUENCE is an INTEGER, not a
single byte/octet.
Secondly, in PER, I remeber I found a similar issue (I had to do a bunch
of workarounds to make the structures always "fill" the octets like the
OPS1955 L7 Ethernet interface expected.