Fixed Serial .read() timeout bugs
The .read() timeout value is set as a Serial object attribute.
We now set it to T1 (T1>T2) at initialization and update it dynamically
during execution for the serial_instance (which is shared resource,
handled by the two classes at the same time).
This is highly unsafe. Thread-safety needs to be a constant concern.