Decreased BAC L2 IF read timeouts
It was indeed a sniffing issue. Either Termite or com0com were
considerably slowing down communication speed.
But indeed, OPS1955 seems to be slower to respond than the TGBV beacon.
Maybe that is due to some post-processing or internal conversion in the
Kapsch device driver. After all, the OPS1955 uses a USB port, and the
BAC L2 port is only a virtual com0com port on the Host computer.