It is best to use OPS1955 driver for testing
The Kapsch OPS1955 Layer7 driver a Windows service that provides a BAC
L7 IF for the beacon at the "OPSCOM" virtual serial port via a Windows
service. So even when no beacon device/HW is connected, we can still
communicate with the service! This is great for quick tests.

The OPSCOM virtual port is connected through a com0com null modem to
COM13. This was done automatically by the OPS1955 L7 service installer.
I am also emulating a COM15 <> COM14 null modem with com0com.
This way I can snoop on the serial line by snooping on COM14 and
forwarding it to COM13! I then connect the HOST to COM15.