Implemented async handling of messages on BAC L2
We use a response queue, where responses await to be obtained by their
requests. This seems inverted. We could instead use Futures, but what
if we send multiple requests of the same type? If we expect multiple
responses, we need a Queue. Normally we never get to send 2 commands of
the same type due to the beacon's timeout constraints. Also, BEACON
responses have priority over HOST requests (ENQ contention).

But for Command ID 0x03 (BST/VST), repeated requests only update the
BeaconID, with only 1 response (VST) being sent upon completion.
So maybe we should just have 1 Future by request, which resolves on the
obtention of a response. We would have to decide how to handle
duplicated requests... Either by cancelling one, or giving the same
response to all of them.