Removed BAC L2 response content error code logging
The second byte of a response does not represent an error code for some
commands. We thus leave the handling of such error codes for an upper
BAC layer, BAC L7. BAC L2 simply respects its role and protocol and
switches messages back and forth from the Beacon to the Host.

DSRC error handling is a responsibility of BAC L7 and DSRC L7.