Added comments to clarify out-of-norm TIS security
TIS uses decimal PAN and ISO 3166 Numeric-3 country codes instead of
Alpha-2 country codes encoded with Baudot ITA2 5-bit characters.
TIS also treats the decimal value of the IssuerId as an hexadecimal
value.
This makes it so the hexadecimal representation of the plaintext for the
derived keys only contains decimal characters (0-9).