Removed unused imports from DSRC L7 RSE module
The threading library is no longer necessary/imported!
We simply use Python's async/await syntax to define async functions.
These return Coroutines, which are Awaitables.

We also removed the asyncio library import.
The event loop is not part of the DSRC L7 backend. Some higher layer
(Application layer) should handle the event loop and the execution of
the corresponding async tasks and subprocesses it runs.