Updated IT_CEN TD name in config.
Italian DSRC (UNI) follows the ETSI-EFC norm, not ISO 14906.
As such, it is an entirely different EFC application, and thus not
compatible with the DSRC L7 communication of this project.

IT_CEN is simply a "fake" Toll Domain, where some OBUs support ISO 14906
DSRC communication with the Italian instance in the OBU's memory.
But it is not a true, actual Toll Domain...